This commit was manufactured by cvs2svn to create tag 'merged_to_RC_1_30_0'.
[SVN r19627]

+++ pyste/tests/example_templatesUT.py
@@ -6,19 +6,19 @@ class TemplatesTest(unittest.TestCase):
     def testIt(self):  
         fp = FPoint()
         fp.i = 3.0
-        fp.j = 4.0
+        fp.j = 4
         ip = IPoint()
         ip.x = 10
-        ip.y = 3
+        ip.y = 3.0
 
         self.assertEqual(fp.i, 3.0)
-        self.assertEqual(fp.j, 4.0)
+        self.assertEqual(fp.j, 4)
         self.assertEqual(ip.x, 10)
-        self.assertEqual(ip.y, 3)
+        self.assertEqual(ip.y, 3.0)
         self.assertEqual(type(fp.i), float)
-        self.assertEqual(type(fp.j), float)
+        self.assertEqual(type(fp.j), int)
         self.assertEqual(type(ip.x), int)
-        self.assertEqual(type(ip.y), int)
+        self.assertEqual(type(ip.y), float)
